requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tensorflow>=2.6.0
pyspark>=3.1.0
kafka-python>=2.0.2
paho-mqtt>=1.6.1
